+++ infra/resources/_modules/testing/locals.tf
@@ -14,7 +14,7 @@ locals {
   subnet_names_by_id = { for subnet_id in var.vnet_common.subnet_ids : subnet_id => provider::azurerm::parse_resource_id(subnet_id)["resource_name"] }
   subnets_to_associate = {
     for subnet_id, subnet_name in local.subnet_names_by_id : subnet_id => subnet_name
-    if !contains([var.runner_subnet_name, "GatewaySubnet"], subnet_name)
+    if !contains([var.runner_subnet_name, "GatewaySubnet"], subnet_name) && !strcontains(subnet_name, "apim")
   }
 
   vnet_address_space_by_mode = { for mode_key, mode_value in var.test_modes : mode_key => format("10.%d.0.0/16", 20 + index(tolist(var.test_modes), mode_key)) }