+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pagopa/eslint-config
 
+## 5.1.1
+
+### Patch Changes
+
+- 9d7123a: Update dependencies
+
 ## 5.1.0
 
 ### Minor Changes